pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pro-mgr"
version = "0.1.0"
description = "A command-line tool to manage multiple projects from anywhere"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Pro-Mgr Team"}
]
keywords = ["cli", "project-management", "developer-tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "textual>=0.40.0",
    "watchdog>=3.0.0",
    "gitpython>=3.1.0",
    "toml>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
pro-mgr = "pro_mgr.cli:cli"

[project.urls]
Homepage = "https://github.com/username/pro-mgr"
Repository = "https://github.com/username/pro-mgr.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
